Fold SO3LR physics into shared ZBL, electrostatics, and dispersion modules.
Delete so3lr_physics.py; make ElectrostaticEnergy flavor=SO3LR for erf Coulomb,
rewrite ZBL with correct eV ke default and optional switch_off, and add
TSQDODispersion under layers/dispersion/.
