+++ docs/developer_guide.rst
@@ -412,7 +412,7 @@ Frank et al. (NeurIPS 2022) lives under :code:`enerzyme/models/so3krates/` with 
 SO3LR (:code:`so3lr`)
 ^^^^^^^^^^^^^^^^^^^^
 
-Kabylda et al. (JACS 2025) is registered as :code:`architecture: so3lr` but **reuses** :code:`So3kratesCore`. Defaults and layers live in :code:`enerzyme/models/so3krates/so3lr.py`. New post-core priors (:code:`SO3LRZBLRepulsionEnergy`, :code:`ErfCoulombEnergy`, :code:`TSQDODispersionEnergy`) and heads (:code:`PartialChargeReadout`, :code:`HirshfeldReadout`, :code:`ChargeSpinEmbedding`) are architecture-agnostic under :code:`enerzyme/models/layers/`. SpookyNet/PhysNet ZBL, shielded Coulomb, and Grimme D3/D4 are left unchanged — SO3LR physics differs (1.5 Å ZBL switch, ``erf`` Coulomb, Hirshfeld TS–QDO). Cutoff alias :code:`phys` → polynomial. Tests: :code:`test/test_so3lr.py`. Enerzymette: :code:`architecture: so3lr` + resolved :code:`config.yaml` (see :code:`enerzyme/config/so3lr_layers_example.yaml`).
+Kabylda et al. (JACS 2025) is registered as :code:`architecture: so3lr` but **reuses** :code:`So3kratesCore`. Defaults and layers live in :code:`enerzyme/models/so3krates/so3lr.py`. Physics uses shared modules with SO3LR options: :code:`ZBLRepulsionEnergy` (:code:`switch_off`), :code:`ElectrostaticEnergy` (:code:`flavor: SO3LR`), :code:`TSQDODispersionEnergy` under :code:`enerzyme/models/layers/dispersion/`, plus :code:`PartialChargeReadout` / :code:`HirshfeldReadout` / :code:`ChargeSpinEmbedding`. Grimme D3/D4 remain for PhysNet/SpookyNet stacks. Cutoff alias :code:`phys` → polynomial. Tests: :code:`test/test_so3lr.py`. Enerzymette: :code:`architecture: so3lr` + resolved :code:`config.yaml` (see :code:`enerzyme/config/so3lr_layers_example.yaml`).
 
 AllScAIP (:code:`AllScAIP`)
 ^^^^^^^^^^^^^^^^^^^^^^^^^
