+++ enerzyme/models/layers/dispersion/__init__.py
@@ -1,2 +1,3 @@
 from .grimme_d3 import GrimmeD3EnergyLayer
 from .grimme_d4 import GrimmeD4EnergyLayer
+from .ts_qdo import TSQDODispersionEnergyLayer
